+++ ci/scripts/r_build.sh
@@ -20,14 +20,17 @@ set -ex
 
 : ${R_BIN:=R}
 source_dir=${1}/r
-with_docs=${2:-false}
+build_dir=${2}
+
+: ${BUILD_DOCS_R:=OFF}
 
 pushd ${source_dir}
 
 ${R_BIN} CMD INSTALL ${INSTALL_ARGS} .
 
-if [ "${with_docs}" == "true" ]; then
+if [ "${BUILD_DOCS_R}" == "ON" ]; then
   ${R_BIN} -e "pkgdown::build_site(install = FALSE)"
+  rsync -a ${source_dir}/docs/ ${build_dir}/docs/r
 fi
 
 popd