+++ cpp/src/arrow/csv/options.h
@@ -180,6 +180,9 @@ struct ARROW_EXPORT WriteOptions {
   /// This number can impact performance.
   int32_t batch_size = 1024;
 
+  /// \brief The string to write for null values. Quotes are not allowed in this string.
+  std::string null_string;
+
   /// \brief IO context for writing.
   io::IOContext io_context;